two_eyes_vision_github/requirements.txt
opencv-python
numpy
pyserial
matplotlib
